api/requirements.txt
fastapi
uvicorn
supabase
pandas
scikit-learn
numpy
python-dotenv